+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- fix imports to point to fork version `github.com/helstern/kacl` instead of `github.com/nicwest/kacl`
+
 ## [v0.2.0] - 2020-01-25
 ### Changed
 - Allow versions numbers with the format `vX.Y.Z`